airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow
apache-airflow-providers-google
apache-airflow-providers-postgres
astronomer-cosmos
dbt-core
dbt-bigquery